Split test to different files
