+++ boards/tests/test_view_new_topic.py
@@ -1,87 +1,10 @@
 from django.test import TestCase
-from django.urls import reverse, resolve
 from django.contrib.auth.models import User
+from django.urls import reverse, resolve
 
-from ..views import home, board_topics, new_topic
+from ..views import new_topic
 from ..models import Board, Topic, Post
 from ..forms import NewTopicForm
-
-
-class HomeTests(TestCase):
-    """ Defines tests for the home page
-    """
-
-    def setUp(self):
-        self.board = Board.objects.create(name='Django', description='Django board.')
-        url = reverse('home')
-        self.response = self.client.get(url)
-
-    def test_home_view_status_code(self):
-        """Test the home view status code.
-        """
-        self.assertEquals(self.response.status_code, 200)
-
-    def test_home_url_resolves_home_view(self):
-        """Test the Home Url resolves '/'.
-        """
-
-        view = resolve('/')
-        self.assertEquals(view.func, home)
-
-    def test_home_view_contains_link_to_topics_page(self):
-        """Test Home view contains Link to topics page.
-        """
-        board_topics_url = reverse('board_topics', kwargs={'pk': self.board.pk})
-        self.assertContains(self.response, 'href="{0}"'.format(board_topics_url))
-
-
-class BoardTopicsTests(TestCase):
-    """Test for the Board Topics relationship.
-    """
-
-    def setUp(self):
-        Board.objects.create(name='Django', description='Django board.')
-
-    def test_board_topics_view_success_status_code(self):
-        """Test Board topics view success status.
-        """
-        url = reverse('board_topics', kwargs={'pk': 1})
-        response = self.client.get(url)
-        self.assertEquals(response.status_code, 200)
-
-    def test_board_topics_view_not_found_status_code(self):
-        """Test board  topics view not found status code.
-        """
-
-        url = reverse('board_topics', kwargs={'pk': 99})
-        response = self.client.get(url)
-        self.assertEquals(response.status_code, 404)
-
-    def test_board_topics_url_resolves_board_topics_view(self):
-        """Test Board topics Url resolves to boards topics view.
-        """
-        view = resolve('/boards/1/')
-        self.assertEquals(view.func, board_topics)
-
-    def test_board_topics_view_contains_link_back_to_homepage(self):
-        """Test Board topics view contains link to homepage.
-        """
-        board_topics_url = reverse('board_topics', kwargs={'pk': 1})
-        response = self.client.get(board_topics_url)
-        homepage_url = reverse('home')
-        self.assertContains(response, 'href="{0}"'.format(homepage_url))
-
-    def test_board_topics_view_contains_navigation_links(self):
-        """Test board topics view contains navigation links.
-        """
-        board_topics_url = reverse('board_topics', kwargs={'pk': 1})
-        homepage_url = reverse('home')
-        new_topic_url = reverse('new_topic', kwargs={'pk': 1})
-
-        response = self.client.get(board_topics_url)
-
-        self.assertContains(response, 'href="{0}"'.format(homepage_url))
-        self.assertContains(response, 'href="{0}"'.format(new_topic_url))
 
 
 class NewTopicTests(TestCase):
